If spec A has more than a one variable that does not appear in the other spec, we unfortunately have to adapt the composition and create more restsOfTheUniverse.
The next-state relation of this composition evaluates to false after the
initial state, because it defines variables z and zz to always have
equal values.  This is only satisfied in the special case where the
values of variables z and zz in A are equal throughout all behaviors
(e.g. Init changed to ... /\ zz = TRUE).